added column summation for encrypted matrix and sigmoid evaluation

+++ util/bench.py
@@ -25,9 +25,9 @@ def bench_model(model, num_runs:int, n:int, m:int, k:int) -> dict:
 
 def gen_random_data(seed: int, n: int, m: int, k: int) -> tuple():
     np.random.seed(seed)
-    S = np.random.normal(0, 1, size=(n, m)) # SNP matrix
+    S = np.random.normal(0, 1, size=(n, m)).astype(np.float64) # SNP matrix
     X0 = np.random.uniform(0, 1, size=(n, k)) # covariates matrix 
     ones = np.zeros(n) + 1
-    X = np.append(ones.reshape(n, 1), X0, axis=1) # covariates matrix concatenated with const term in regression
-    y = np.random.binomial(1, 0.5, n).reshape(n,1)
+    X = np.append(ones.reshape(n, 1), X0, axis=1).astype(np.float64) # covariates matrix concatenated with const term in regression
+    y = np.random.binomial(1, 0.5, n).reshape(n,1).astype(np.float64)
     return (y,S,X)
